refactor(memory): own the frontmatter parser (narrow skills coupling)
Part of narrowing memory's coupling to the host skill subsystem before any
@vellumai/plugin-api facet (Sidd's "narrow the coupling first" call): the
YAML frontmatter parser memory used from skills/frontmatter is a generic
parser, not a capability-catalog read, so memory gets its own copy
(plugins/defaults/memory/frontmatter.ts) and stops importing the host module.

- New plugins/defaults/memory/frontmatter.ts: FRONTMATTER_REGEX +
  parseFrontmatterFields (verbatim; imports yaml + getLogger).
- Repoint the 5 memory importers (v2/page-store, v2/frontmatter-sweep,
  v3/card, v3/edge, v3-eval/eval-packets) to the local copy.
- Host skills/frontmatter stays (config/skills + prompts/sections still use it).
- Boundary baseline drops skills/frontmatter from the memory key.

Zero @vellumai/plugin-api change; behavior byte-identical (verbatim parser).
