+++ assistant/src/plugins/defaults/memory/v3/card.ts
@@ -1,7 +1,4 @@
-import {
-  FRONTMATTER_REGEX,
-  parseFrontmatterFields,
-} from "../../../../skills/frontmatter.js";
+import { FRONTMATTER_REGEX, parseFrontmatterFields } from "../frontmatter.js";
 import { injectedConceptHeader } from "../v2/injected-block-slugs.js";
 import type { Slug } from "./types.js";
